+++ changelog.md
@@ -3,7 +3,9 @@
 #### v1.1.0-beta (2023-12-xx)
 
 This release focuses on supporting runes introduced in Season of Discovery
+- New SAO: Druid's Fury of Stormrage
 - New SAO: Mage's Fingers of Frost
+- New GAB: Druid's Healing Touch, during Fury of Stormrage
 - New GAB: Mage's Ice Lance, during Fingers of Frost
 - New GAB: Mage's Ice Lance, when the enemy target is Frozen